stream_dad/requirements.txt
# Core dependencies
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Data processing
PyYAML>=6.0
h5py>=3.1.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Logging and experiment tracking
wandb>=0.12.0
tensorboard>=2.8.0

# Optional: for better performance
numba>=0.56.0

# Development and testing
pytest>=6.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Utilities
tqdm>=4.62.0
pathlib>=1.0.0
argparse>=1.4.0